requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
greenlet==1.1.0
idna==2.10
passlib==1.7.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
SQLAlchemy==1.4.15
sqlparse==0.4.1
stripe==2.57.0
urllib3==1.26.5
